Fixed mean calculation. Was not converting to double correctly.
Former-commit-id: 3fe53f8770b366478f6b399f9608297b552b8e66
Former-commit-id: 8ad9c395c75b80cfc46936ba3417b0ad80ed2584

+++ NINA/Model/MyCamera/ImageArray.cs
@@ -90,7 +90,7 @@ namespace NINA.Model.MyCamera {
                     oldmax = max;
                 }
 
-                double mean = sum / count;
+                double mean = sum / (double)count;
                 double variance = (squareSum - count * mean * mean) / (count);
                 double stdev = Math.Sqrt(variance);
